Scene Replicated
While I would love to sue a premade scene, the one I had already was compelted.
But after making this I couldnt find my way into the Navegation tab. I cannot find it and I have little time to do it now, so it will have to be resolved in class with assistance.